renamed homepage.js to intro.js
